+++ client/src/App.tsx
@@ -1,0 +1,12 @@
+import React from 'react'
+import SDDDashboard from './components/SDDDashboard'
+
+function App() {
+  return (
+    <div className="min-h-screen bg-gray-900">
+      <SDDDashboard />
+    </div>
+  )
+}
+
+export default App